Remove skull translation events and define custom blocks for custom skulls
Clean up skull block translation a bit

+++ api/geyser/src/main/java/org/geysermc/geyser/api/block/custom/component/RotationComponent.java
@@ -25,5 +25,5 @@
 
 package org.geysermc.geyser.api.block.custom.component;
 
-public record RotationComponent(float x, float y, float z) {
+public record RotationComponent(int x, int y, int z) {
 }
